backend/requirements.txt
fastapi
uvicorn
motor
google-generativeai
python-multipart
python-dotenv
email-validator
pymongo
bcrypt
python-jose[cryptography]
fastapi-mail
cloudinary
python-multipart
